Finished disassembly of server.js into comment.js.
